[FEAT]#6: 카드 관련 기능 리팩토링
