+++ src/main/java/com/umc/cardify/service/CardService.java
@@ -1,5 +1,10 @@
 package com.umc.cardify.service;
 
+import com.umc.cardify.domain.Card;
+import com.umc.cardify.domain.Note;
+import com.umc.cardify.domain.enums.Side;
+import com.umc.cardify.dto.note.NoteRequest;
+import com.umc.cardify.repository.CardRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -23,6 +28,7 @@ public class CardService {
 
 	private final OverlayRepository overlayRepository;
 
+	private final CardRepository cardRepository;
 	public String addImageCard(Long userId, MultipartFile image, CardRequest.addImageCard request) {
 		String imgUrl = s3Service.upload(image, "imageCards");
 
@@ -50,5 +56,32 @@ public class CardService {
 		}
 
 		return savedImageCard.getImageUrl();
+	}
+	public void addCard(CardRequest.WriteCardDto cardDto, Note note){
+		String contents_front = cardDto.getText();
+		String contents_back = contents_front
+				.replace(">>", "")
+				.replace("<<", "")
+				.replace("{{", "")
+				.replace("}}", "")
+				.replace("==", "");
+
+		Card card_front = Card.builder()
+				.note(note)
+				.name(cardDto.getName())
+				.contents(contents_front)
+				.side(Side.FRONT)
+				.countLearn(0L)
+				.build();
+		Card card_back = Card.builder()
+				.note(note)
+				.name(cardDto.getName())
+				.contents(contents_back)
+				.side(Side.BACK)
+				.countLearn(0L)
+				.build();
+
+		cardRepository.save(card_front);
+		cardRepository.save(card_back);
 	}
 }
