+++ src/main/java/com/umc/cardify/service/NoteService.java
@@ -9,6 +9,7 @@ import com.umc.cardify.domain.Note;
 import com.umc.cardify.domain.User;
 import com.umc.cardify.domain.enums.MarkStatus;
 import com.umc.cardify.domain.enums.Side;
+import com.umc.cardify.dto.card.CardRequest;
 import com.umc.cardify.dto.note.NoteRequest;
 import com.umc.cardify.dto.note.NoteResponse;
 import com.umc.cardify.repository.CardRepository;
@@ -33,6 +34,7 @@ public class NoteService {
     private final UserRepository userRepository;
     private final NoteConverter noteConverter;
     private final CardRepository cardRepository;
+    private final CardService cardService;
     public Note getNoteToID(long noteId){
         return noteRepository.findById(noteId).orElseThrow(()-> new BadRequestException(ErrorResponseStatus.NOT_FOUND_ERROR));
     }
@@ -137,42 +139,15 @@ public class NoteService {
             note.setContents(request.getContents());
             noteRepository.save(note);
 
-            List<NoteRequest.WriteCardDto> cardsDto = request.getCards();
+            List<CardRequest.WriteCardDto> cardsDto = request.getCards();
             if(note.getCards() != null) {
                 List<Card> cardList = cardRepository.findByNote(note);
                 cardRepository.deleteAll(cardList);
             }
             if (cardsDto != null) {
-                cardsDto.forEach((card) -> addCard(card, note));
+                cardsDto.forEach((card) -> cardService.addCard(card, note));
             }
             return true;
         }
-    }
-    public void addCard(NoteRequest.WriteCardDto cardDto, Note note){
-        String contents_front = cardDto.getText();
-        String contents_back = contents_front
-                .replace(">>", "")
-                .replace("<<", "")
-                .replace("{{", "")
-                .replace("}}", "")
-                .replace("==", "");
-
-        Card card_front = Card.builder()
-                .note(note)
-                .name(cardDto.getName())
-                .contents(contents_front)
-                .side(Side.FRONT)
-                .countLearn(0L)
-                .build();
-        Card card_back = Card.builder()
-                .note(note)
-                .name(cardDto.getName())
-                .contents(contents_back)
-                .side(Side.BACK)
-                .countLearn(0L)
-                .build();
-
-        cardRepository.save(card_front);
-        cardRepository.save(card_back);
     }
 }
